refactor(data-types): move web-search vocabulary below preference
The provider vocabulary (ids, capabilities, overrides, WebSearchProvider)
lived in preferenceTypes.ts, so the types layer imported preference —
upside down. The vocabulary now lives in types/webSearch and preference
imports it for its key value types; consumers import it from its owner
instead of through the preference barrel.
