+++ src/backend/services/webSearch/providers/api/__tests__/FirecrawlProvider.test.ts
@@ -1,5 +1,4 @@
-import type { WebSearchProvider } from '@/shared/data/preference';
-import type { WebSearchExecutionConfig } from '@/shared/data/types/webSearch';
+import type { WebSearchProvider, WebSearchExecutionConfig } from '@/shared/data/types/webSearch';
 
 import { ApiKeyRotationState } from '../../../utils/provider';
 import firecrawlResponse from '../../__tests__/fixtures/firecrawl-response.json';
