+++ src/backend/services/webSearch/providers/api/__tests__/JinaProvider.test.ts
@@ -1,6 +1,5 @@
 import { ApiKeyRotationState } from '@/backend/services/webSearch/utils/provider';
-import type { WebSearchProvider } from '@/shared/data/preference';
-import type { WebSearchExecutionConfig } from '@/shared/data/types/webSearch';
+import type { WebSearchProvider, WebSearchExecutionConfig } from '@/shared/data/types/webSearch';
 
 import { JinaProvider } from '../JinaProvider';
