+++ src/backend/services/webSearch/providers/api/__tests__/SearxngProvider.test.ts
@@ -1,5 +1,4 @@
-import type { WebSearchProvider } from '@/shared/data/preference';
-import type { WebSearchExecutionConfig } from '@/shared/data/types/webSearch';
+import type { WebSearchProvider, WebSearchExecutionConfig } from '@/shared/data/types/webSearch';
 
 import { ApiKeyRotationState } from '../../../utils/provider';
 import searxngSearchResponse from '../../__tests__/fixtures/searxng-search-response.json';
