+++ src/frontend/features/settings/WebSearchScreen/utils/providerSettings.ts
@@ -1,14 +1,14 @@
-import type {
-  WebSearchCapability,
-  WebSearchProviderId,
-  WebSearchProviderOverride,
-  WebSearchProviderOverrides,
-} from '@/shared/data/preference';
 import {
   MOBILE_SUPPORTED_WEB_SEARCH_PROVIDERS,
   WEB_SEARCH_PROVIDER_PRESET_MAP,
   type WebSearchProviderPreset,
 } from '@/shared/data/presets/webSearchProviders';
+import type {
+  WebSearchCapability,
+  WebSearchProviderId,
+  WebSearchProviderOverride,
+  WebSearchProviderOverrides,
+} from '@/shared/data/types/webSearch';
 
 export type WebSearchProviderCapability = WebSearchProviderPreset['capabilities'][number];
