+++ src/shared/data/types/webSearch.ts
@@ -1,10 +1,86 @@
-import type {
-  WebSearchCapability,
-  WebSearchCompressionMethod,
-  WebSearchProvider,
-  WebSearchProviderId,
-  WebSearchProviderOverrides,
-} from '@/shared/data/preference';
+// ============================================================================
+// Provider Vocabulary
+// ============================================================================
+
+export const WEB_SEARCH_PROVIDER_TYPES = ['api', 'mcp'] as const;
+
+export type WebSearchProviderType = (typeof WEB_SEARCH_PROVIDER_TYPES)[number];
+
+export const WEB_SEARCH_PROVIDER_IDS = [
+  'zhipu',
+  'tavily',
+  'searxng',
+  'exa',
+  'exa-mcp',
+  'bocha',
+  'querit',
+  'fetch',
+  'jina',
+  'firecrawl',
+] as const;
+
+export type WebSearchProviderId = (typeof WEB_SEARCH_PROVIDER_IDS)[number];
+
+export const WEB_SEARCH_CAPABILITIES = ['searchKeywords', 'fetchUrls'] as const;
+
+export type WebSearchCapability = (typeof WEB_SEARCH_CAPABILITIES)[number];
+
+export type WebSearchProviderCapabilityOverride = {
+  apiHost?: string;
+};
+
+export type WebSearchProviderCapabilityOverrides = Partial<
+  Record<WebSearchCapability, WebSearchProviderCapabilityOverride>
+>;
+
+export type WebSearchProviderOverride = {
+  apiKeys?: string[];
+  capabilities?: WebSearchProviderCapabilityOverrides;
+  engines?: string[];
+  basicAuthUsername?: string;
+  basicAuthPassword?: string;
+};
+
+export type WebSearchProviderOverrides = Partial<
+  Record<WebSearchProviderId, WebSearchProviderOverride>
+>;
+
+/**
+ * Full WebSearch Provider configuration
+ * Generated at runtime by merging preset with user overrides
+ */
+export interface WebSearchProvider {
+  /** Unique provider identifier */
+  id: WebSearchProviderId;
+  /** Display name (from preset) */
+  name: string;
+  /** Provider type (from preset) */
+  type: WebSearchProviderType;
+  /** API keys (from user overrides) */
+  apiKeys: string[];
+  /** Capability API settings (user override merged into preset capabilities) */
+  capabilities: {
+    feature: WebSearchCapability;
+    /** Can be empty for self-hosted or hostless providers; resolve and validate via resolveProviderApiHost. */
+    apiHost?: string;
+  }[];
+  /** Search engines (from user overrides) */
+  engines: string[];
+  /** Basic auth username (from user overrides) */
+  basicAuthUsername: string;
+  /** Basic auth password (from user overrides) */
+  basicAuthPassword: string;
+}
+
+/**
+ * Compression method type
+ * Stored in chat.web_search.compression.method
+ */
+export type WebSearchCompressionMethod = 'none' | 'cutoff';
+
+// ============================================================================
+// Runtime Config And Wire Types
+// ============================================================================
 
 export const DEFAULT_WEB_SEARCH_CUTOFF_LIMIT = 2000;
